Adding %use-par-map, a parameter
Note:  Guile-CV extensively use par-map.  However, there are two
situations it should not: (1) when a procedure calls other Guile-CV
procedures; (2) because the user already calls par-map, then calls
Guile-CV procedures.  There is actually a third case, which is when
images are small, since there is a cost for multi threading, it might be
interesting to parameterize %use-par-map to #f, up to users to decide.

* cv/adds.scm:
* cv/filters.scm:
* cv/idata.scm:
* cv/imgproc.scm:
* cv/init.scm:
* cv/morphology.scm:
* cv/segmentation.scm:  All calls to par-map are now subject to both
  the number of channels and the value returned by (%use-par-map).