One f32vector API design change, three new procedures
I should have split this commit in two, but I realized it too late. The
API design change is that we should use #:key, not #:optional, unless
you are 100% sure the API for the procedures using #:optional will never
change: at this still early stage of design and development, we are
likely to change things.

* cv/support/f32vector.scm:  (a) adding three new procedure:
  f32vector-reduce, f32vector-mean, f32vector-std-dev; (b) use #:key,
  not #:optional.

* test-suite/tests/support.scm:  New tests for f32vector-reduce,
  f32vector-mean and f32vector-std-dev.

* cv/idata.scm:  Updated to use #:prec keyword new API for f32vector-list=?