Introducing latex dependency checks
Notes: (a)im-histogram uses latex to produce the header, the legend
indices and the footer of its returned histogram image.  Later we might
use pangocairo, bit right now I'm not sure I want to make guile-cv
depends on guile-gnome, even for one module only (gnome pangocairo).
Till g-golf is working, could also just bind paingocairo, but anyway,
right now latex is fine;

(b) In order for the result to be beautifull, at least that'd what I
think, I have decided to use the Iwona math font, a free font of course,
here: http://www.tug.dk/FontCatalogue/iwona, not always distributed with
the 'basic' latex distro - on Debian for example, it is part of the
texlive-fonts-extra package - but easy to grab and install.

* m4/latex.m4:
* build-aux/latex.scm:  New files.

* configure.ac:  Updated to check that the latex standalone documentclass
  is available, as well as the following packages: inputenc, fontenc,
  lmodern, xcolor, booktabs, siunitx, iwona.

* .gitignore:
* Makefile.am:  Updated for the above.