Performance improvements - im-particle-clean
Note: this commit also includes a few C functions that I wrote to see if
it could improve bb-intersect?, but it actually got things slower. Guile
scheme small integer operations are really fast already, and there is an
overhead in calling C functons.  Nevertheless, I decided to keep the
code as an example.

* cv/adds.scm (im-particle-clean): Calls f32vector-scrap-in-place, itself
  calling the newly added C function.

* cv/support/f32vector.scm (f32vector-scrap-in-place):
* cv/support/libguile-cv.scm (f32vector-scrap-in-place-c):  New procedures.

* libguile-cv/libguile-cv.h:
* libguile-cv/libguile-cv.c (f32vector_scrap_in_place_c):  New function.

* NEWS:  Updated for the above.