Moving im-transpose core functionality to libguile-cv
Note that all memory allocation - except for some local variables
- still is being performed in scheme.

This move makes im-delineate work almost twice as fast. Because
all memory allocation still is done in scheme, and not in C, I am
very confortable with this approach. When Guile has an AOT
compiler that compiles fixed size floating points and small
integers operations and vector loops for these that runs as fast
as in C, and I am prety confident that it will happen, maybe not
tomorrow, but it will happen, we will revert these and use Guile
Scheme again.

* libguile-cv/libguile-cv.c:
* libguile-cv/libguile-cv.h (f32vector_transpose_c):  New function.

* cv/adds.scm:
* cv/support/f32vector.scm:
* cv/support/libguile-cv.scm:  Updated for the above.

* NEWS:  Updating 'Performance improvements' for the next release.